# apps/api/docs/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Yet Another Kanban
  description: Yet another kanban board API
  version: v1
paths:
...
  '/Api/Workspaces/{workspaceId}/RenameGroup/{groupId}':
    put:
      tags:
        - Groups
      summary: Rename specified group
      operationId: RenameGroup
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/RenameGroupRequest'
          text/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/RenameGroupRequest'
          application/*+json:
            schema:
              allOf:
                - $ref: '#/components/schemas/RenameGroupRequest'
      responses:
        '200':
          description: Success
...
components:
  schemas:
...
    RenameGroupRequest:
      required:
        - name
      type: object
      properties:
        name:
          type: string
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 12345abcdef\""
      name: Authorization
      in: header
security:
  - Bearer: [ ]